tidy(common): Split off slogger from common  (#97)
This is another split of the common library to help encapsulate and
reduce size/usage of the common library, for both new and future
libraries.
