+++ CMakeLists.txt
@@ -362,6 +362,7 @@ endif()
 add_subdirectory(external)
 add_subdirectory(src/admin)
 add_subdirectory(src/errors)
+add_subdirectory(src/slogger)
 add_subdirectory(src/common)
 add_subdirectory(src/protocol)
 add_subdirectory(src/mount)
