+++ src/chunkserver/network_main_thread.cc
@@ -57,7 +57,7 @@
 #include "common/massert.h"
 #include "protocol/SFSCommunication.h"
 #include "protocol/packet.h"
-#include "common/slogger.h"
+#include "slogger/slogger.h"
 #include "common/sockets.h"
 #include "devtools/TracePrinter.h"
