+++ src/master/personality.cc
@@ -28,7 +28,7 @@
 #include "common/event_loop.h"
 #include "common/main.h"
 #include "common/massert.h"
-#include "common/slogger.h"
+#include "slogger/slogger.h"
 
 namespace metadataserver {
