+++ src/master/topology.cc
@@ -34,7 +34,7 @@
 #include "common/event_loop.h"
 #include "common/massert.h"
 #include "errors/sfserr.h"
-#include "common/slogger.h"
+#include "slogger/slogger.h"
 #include "master/itree.h"
 
 static void *racktree;
