+++ src/metarestore/merger.cc
@@ -29,7 +29,7 @@
 
 #include "protocol/SFSCommunication.h"
 #include "errors/saunafs_error_codes.h"
-#include "common/slogger.h"
+#include "slogger/slogger.h"
 #include "master/restore.h"
 
 #define BSIZE 200000
