+++ src/mount/fuse/sfs_meta_fuse.cc
@@ -31,7 +31,7 @@
 
 #include "common/datapack.h"
 #include "protocol/SFSCommunication.h"
-#include "common/slogger.h"
+#include "slogger/slogger.h"
 #include "common/special_inode_defs.h"
 #include "mount/exports.h"
 #include "mount/mastercomm.h"
