+++ src/slogger/CMakeLists.txt
@@ -1,0 +1,5 @@
+collect_sources(SLOGGER)
+
+shared_add_library(slogger ${SLOGGER_SOURCES})
+target_link_libraries(slogger sfserr)
+target_link_libraries(slogger_pic sfserr_pic)
